+++ CHANGELOG.md
@@ -12,7 +12,7 @@
         - ENSIntegration, IexecOrderManagement, IexecRelay (#195)
         - IexecCategoryManager, IexecERC20 (#192)
         - test/*fullchain* (#190)
-        - IexecAccessors (#189, #191)
+        - IexecAccessors, IexecMaintenance (#189, #191)
 - Migrate scripts to TypeScript: (#184)
     - `getFunctionSignatures.js`, `common-test-snapshot.js`, `test-storage.js`, `timelock.js`
 - Migrated utility files to TypeScript : (#183)
